pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mtgslider"
version = "0.1.0"
description = "MTG thematic research and slideshow pipeline"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "python-pptx>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
mtgslider = "mtgslider.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
